+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   * Array index section: `(.[i])` is shorthand for `(\x -> x[i])`.
     Full slice syntax supported. (#559)
 
+  * New `assert` construct. (#464)
+
 ### Removed
 
 ### Changed